refactor: splash_screen 파일 위치 이동
- presentation/wedgets/ 에서 presentation/ 으로 이동
