+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:weve_client/commons/presentation/main_screen.dart';
-import 'package:weve_client/commons/presentation/widgets/splash_screen.dart';
+import 'package:weve_client/commons/presentation/splash_screen.dart';
 import 'package:weve_client/core/constants/colors.dart';
 
 void main() {
